+++ scene.rt
@@ -5,4 +5,4 @@ L   -40,0,30                    0.7             255,255,255
 
 pl  0,0,0           0,1.0,0                     255,0,255
 sp  0,0,20                      20              255,0,0
-cy  50.0,101,20.6   0,0,1.0     14.2    21.42   10,0,255
+cy  50.0,10.0,20.6   0,0,1.0     14.2    21.42   10,0,255
